+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix `validate` reporting "No ShouldCatch attributes found" instead of surfacing the underlying error when run on a directory outside a git repository (https://github.com/svenvanheugten/mutannot/issues/11)
 
+### Added
+
+- Add a `--jobs` option to `run` to run mutations in parallel.
+
 ## [0.7.1] - 2026-06-26
 
 ### Fixed